[Bugfix] Now the fg PID is checked before sending a kill signal.

	modified:   usr/share/perl5/Zevenet/Farm/L4xNAT/Action.pm
	modified:   usr/share/perl5/Zevenet/Farm/L4xNAT/Backend.pm
	modified:   usr/share/perl5/Zevenet/Farm/L4xNAT/Config.pm

+++ usr/share/perl5/Zevenet/Farm/L4xNAT/Action.pm
@@ -471,7 +471,7 @@ sub setL4NewFarmName    # ($farm_name,$new_farm_name)
 
 	if ( $$farm{ status } eq 'up' )
 	{
-		if ( $fg_enabled eq 'true' )
+		if ( $fg_enabled eq 'true' && $fg_pid > 0 )
 		{
 			kill 'STOP' => $fg_pid;
 		}
@@ -589,7 +589,7 @@ sub setL4NewFarmName    # ($farm_name,$new_farm_name)
 
 		if ( $fg_enabled eq 'true' )
 		{
-			if ( $0 !~ /farmguardian/ )
+			if ( $0 !~ /farmguardian/ && $fg_pid > 0 )
 			{
 				kill 'CONT' => $fg_pid;
 			}
